refactor(playground): move message media out of message_helpers
`message_helpers` describes how a prompt template becomes the message dicts an LLM
client takes, and media touched nearly every stage of it: the block types, four
readers over them, the resolution step that attaches bytes, the variable
substitution. Two hundred and eighty added lines, in a module whose job is to be
legible end to end.

It now lives in `message_media`. What stays is what cannot be delegated:

* `content` widening from `str` to `Union[str, list[ContentBlock]]`, on the TypedDict
  and its factory. A string cannot say where an image sits relative to the text, so
  the block list is the media feature's one unavoidable change to the message shape.
* the accumulator in `prompt_chat_template_to_playground_messages` becoming a block
  list rather than a list of strings joined at the end, for the same reason.
* two delegations, four lines and five, where the image/file branch and the
  content-formatting loop used to be inlined.

Thirty-eight added lines instead of two hundred and eighty-two.

As with `media_parts`, the annotations pointing back at `PlaygroundMessage` are
type-checking only and therefore quoted; unquoted, the module raises NameError on
import, which is how this failed the first time here too.
