+++ src/phoenix/server/api/helpers/message_helpers.py
@@ -7,21 +7,17 @@ should be appended to prompt templates when running experiments.
 """
 
 import json
-from typing import Any, Iterable, Literal, Mapping, Optional, Sequence, TypedDict, Union
+from typing import Any, Iterable, Mapping, Optional, Sequence, TypedDict, Union
 
-from sqlalchemy.ext.asyncio import AsyncSession
-from typing_extensions import Required, TypeAlias, assert_never
+from typing_extensions import Required, assert_never
 
-from phoenix.db.types.media import MediaContent
-from phoenix.db.types.prompts import (
-    PromptChatTemplate,
-    PromptTemplateFormat,
+from phoenix.db.types.prompts import PromptChatTemplate, PromptTemplateFormat
+from phoenix.server.api.helpers.message_media import (
+    ContentBlock,
+    TextContentBlock,
+    format_message_content,
+    media_content_block,
 )
-from phoenix.db.types.media_parts import (
-    media_source,
-)
-from phoenix.server.api.exceptions import BadRequest
-from phoenix.server.api.helpers.media import MediaResolutionError, resolve_media
 from phoenix.server.api.types.ChatCompletionMessageRole import ChatCompletionMessageRole
 from phoenix.utilities.template_formatters import (
     FStringTemplateFormatter,
@@ -45,42 +41,6 @@ class PlaygroundToolCall(TypedDict, total=False):
     function: PlaygroundToolCallFunction
 
 
-class TextContentBlock(TypedDict):
-    """A run of text within a message."""
-
-    type: Literal["text"]
-    text: str
-
-
-class MediaContentBlock(TypedDict, total=False):
-    """
-    Binary media within a message, at one of three stages.
-
-    ``variable`` is set when the prompt names the image rather than storing it; the
-    run supplies the reference. :func:`formatted_messages` substitutes the value
-    into ``url``, exactly as it substitutes text variables.
-
-    ``url`` is the reference and is what gets recorded on the span, so that trace
-    attributes stay small and stable.
-
-    ``data`` and the authoritative ``media_type`` hold what the provider needs, and
-    are populated by :func:`resolve_message_media`.
-    """
-
-    type: Required[Literal["media"]]
-    kind: Required[Literal["image", "file"]]
-    """Which content part this came from, which decides the provider's block shape."""
-    variable: str
-    url: str
-    media_type: str
-    data: bytes
-    file_name: str
-    """The stored name, for the providers that require one to carry a document."""
-
-
-ContentBlock: TypeAlias = Union[TextContentBlock, MediaContentBlock]
-
-
 class PlaygroundMessage(TypedDict, total=False):
     role: Required[ChatCompletionMessageRole]
     content: Required[Union[str, list[ContentBlock]]]
@@ -100,84 +60,6 @@ def create_playground_message(
     if tool_calls is not None:
         msg["tool_calls"] = tool_calls
     return msg
-
-
-def message_text(message: PlaygroundMessage) -> str:
-    """
-    The text of a message, with any media omitted.
-
-    Lets a provider integration that cannot send media keep treating message
-    content as a plain string.
-
-    Args:
-        message: The message to read.
-
-    Returns:
-        The message's text, with multiple text blocks joined by newlines.
-    """
-    content = message.get("content")
-    if content is None:
-        return ""
-    if isinstance(content, str):
-        return content
-    return "\n".join(block["text"] for block in content if block["type"] == "text")
-
-
-def message_media(message: PlaygroundMessage) -> list[MediaContentBlock]:
-    """
-    The media blocks of a message, in the order they appear.
-
-    Args:
-        message: The message to read.
-
-    Returns:
-        The message's media blocks, empty if it carries none.
-    """
-    content = message.get("content")
-    if content is None or isinstance(content, str):
-        return []
-    return [block for block in content if block["type"] == "media"]
-
-
-def content_blocks(message: PlaygroundMessage) -> list[ContentBlock]:
-    """
-    The content of a message as an ordered block list.
-
-    Args:
-        message: The message to read.
-
-    Returns:
-        The message's blocks. String content becomes a single text block; empty
-        string content yields no blocks.
-    """
-    content = message.get("content")
-    if content is None:
-        return []
-    if isinstance(content, str):
-        return [TextContentBlock(type="text", text=content)] if content else []
-    return list(content)
-
-
-def reject_media(messages: Iterable[PlaygroundMessage], *, provider: str) -> None:
-    """
-    Reject messages carrying media, for providers without media support yet.
-
-    Fails loudly rather than sending the provider a prompt with its images
-    silently removed.
-
-    Args:
-        messages: The messages about to be sent.
-        provider: Human-readable provider name, used in the error message.
-
-    Raises:
-        BadRequest: Any message carries a media block.
-    """
-    for message in messages:
-        if message_media(message):
-            raise BadRequest(
-                f"{provider} does not support image content in Phoenix yet. "
-                f"Remove the image from the prompt, or run it against Google."
-            )
 
 
 # Mapping from OpenAI role strings to internal enum values
@@ -377,21 +259,7 @@ def prompt_chat_template_to_playground_messages(
                     elif result is not None:
                         blocks.append(TextContentBlock(type="text", text=json.dumps(result)))
                 elif part.type == "image" or part.type == "file":
-                    source = media_source(part)
-                    blocks.append(
-                        MediaContentBlock(
-                            type="media",
-                            kind=part.type,
-                            url=source.url,
-                            media_type=source.media_type,
-                        )
-                        if isinstance(source, MediaContent)
-                        else MediaContentBlock(
-                            type="media",
-                            kind=part.type,
-                            variable=source.variable,
-                        )
-                    )
+                    blocks.append(media_content_block(part))
                 else:
                     assert_never(part)
 
@@ -404,70 +272,6 @@ def prompt_chat_template_to_playground_messages(
             )
         )
     return messages
-
-
-async def resolve_message_media(
-    session: AsyncSession,
-    messages: Iterable[PlaygroundMessage],
-) -> list[PlaygroundMessage]:
-    """
-    Attach resolved bytes to every media block in the given messages.
-
-    Call once the message list is final — after template formatting and after any
-    per-example messages have been appended — so that a single batch resolves every
-    reference across every message.
-
-    Args:
-        session: Session used to read Phoenix-hosted media.
-        messages: Messages whose media should be resolved.
-
-    Returns:
-        New messages whose image blocks carry ``data`` and the authoritative
-        ``media_type``. Messages without media are passed through unchanged.
-
-    Raises:
-        MediaResolutionError: A reference is malformed or names media that is not
-            present.
-    """
-    message_list = list(messages)
-    for message in message_list:
-        for media_block in message_media(message):
-            if "url" not in media_block:
-                # formatted_messages fills a variable's reference in. Reaching here
-                # means the message never went through it.
-                name = media_block.get("variable", "an image")
-                raise MediaResolutionError(f"No image reference was substituted for '{name}'.")
-    urls = [block["url"] for message in message_list for block in message_media(message)]
-    if not urls:
-        return message_list
-
-    resolved = await resolve_media(session, urls)
-    output: list[PlaygroundMessage] = []
-    for message in message_list:
-        content = message.get("content")
-        if isinstance(content, str) or not content:
-            output.append(message)
-            continue
-        blocks: list[ContentBlock] = []
-        for block in content:
-            if block["type"] != "media":
-                blocks.append(block)
-                continue
-            media = resolved[block["url"]]
-            resolved_block = MediaContentBlock(
-                type="media",
-                kind=block["kind"],
-                url=block["url"],
-                media_type=media.media_type,
-                data=media.content,
-            )
-            if media.file_name is not None:
-                resolved_block["file_name"] = media.file_name
-            if (variable := block.get("variable")) is not None:
-                resolved_block["variable"] = variable
-            blocks.append(resolved_block)
-        output.append({**message, "content": blocks})
-    return output
 
 
 def build_template_variables(
@@ -534,32 +338,11 @@ def formatted_messages(
     template_formatter = _template_formatter(template_format=template_format)
     result: list[PlaygroundMessage] = []
     for msg in messages_list:
-        content = msg.get("content")
-        formatted_content: Union[str, list[ContentBlock]]
-        if isinstance(content, str) or content is None:
-            formatted_content = template_formatter.format(content or "", **template_variables)
-        else:
-            blocks: list[ContentBlock] = []
-            for block in content:
-                if block["type"] == "text":
-                    blocks.append(
-                        TextContentBlock(
-                            type="text",
-                            text=template_formatter.format(block["text"], **template_variables),
-                        )
-                    )
-                elif (variable := block.get("variable")) is not None:
-                    blocks.append(
-                        MediaContentBlock(
-                            type="media",
-                            kind=block["kind"],
-                            variable=variable,
-                            url=_media_variable_value(variable, template_variables),
-                        )
-                    )
-                else:
-                    blocks.append(block)
-            formatted_content = blocks
+        formatted_content = format_message_content(
+            msg.get("content"),
+            template_formatter=template_formatter,
+            template_variables=template_variables,
+        )
         result.append(
             create_playground_message(
                 msg["role"],
@@ -569,31 +352,6 @@ def formatted_messages(
             )
         )
     return result
-
-
-def _media_variable_value(variable: str, template_variables: Mapping[str, Any]) -> str:
-    """
-    The media reference supplied for a media variable.
-
-    Args:
-        variable: The media variable's name.
-        template_variables: The values supplied for this run.
-
-    Returns:
-        The reference to resolve, e.g. ``phoenix://media/<sha256>``.
-
-    Raises:
-        BadRequest: No value was supplied, or the value is not a reference string.
-    """
-    if variable not in template_variables:
-        raise BadRequest(f"No image was supplied for '{variable}'.")
-    value = template_variables[variable]
-    if not isinstance(value, str) or not value.strip():
-        raise BadRequest(
-            f"The value supplied for '{variable}' is not an image reference. "
-            f"Upload an image for it and try again."
-        )
-    return value
 
 
 def _template_formatter(template_format: PromptTemplateFormat) -> TemplateFormatter:
